[SQL] move a few helpers up to GithubSource
