+++ src/main/java/com/ververica/platform/FlinkCommitProgram.java
@@ -1,6 +1,6 @@
 package com.ververica.platform;
 
-import static com.ververica.platform.io.source.GithubCommitSource.EVALUATION_ZONE;
+import static com.ververica.platform.io.source.GithubSource.EVALUATION_ZONE;
 import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
 
 import com.ververica.platform.entities.Commit;
