+++ src/main/java/com/ververica/platform/io/source/GithubSource.java
@@ -2,10 +2,14 @@ package com.ververica.platform.io.source;
 
 import java.io.IOException;
 import java.nio.file.Files;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Date;
 import okhttp3.Cache;
 import okhttp3.OkHttpClient;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.streaming.api.functions.source.RichSourceFunction;
+import org.kohsuke.github.GHUser;
 import org.kohsuke.github.GitHub;
 import org.kohsuke.github.GitHubBuilder;
 import org.kohsuke.github.RateLimitChecker;
@@ -14,6 +18,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public abstract class GithubSource<T> extends RichSourceFunction<T> {
+  public static final ZoneId EVALUATION_ZONE = ZoneId.of("UTC");
   private static final Logger LOG = LoggerFactory.getLogger(GithubSource.class);
 
   protected final String repoName;
@@ -22,6 +27,20 @@ public abstract class GithubSource<T> extends RichSourceFunction<T> {
 
   public GithubSource(String repoName) {
     this.repoName = repoName;
+  }
+
+  protected static LocalDateTime dateToLocalDateTime(Date date) {
+    return date.toInstant().atZone(EVALUATION_ZONE).toLocalDateTime();
+  }
+
+  protected static String getUserName(GHUser user) throws IOException {
+    if (user == null) {
+      return "unknown";
+    } else if (user.getName() == null) {
+      return user.getLogin() == null ? "unknown" : user.getLogin();
+    } else {
+      return user.getName();
+    }
   }
 
   @Override
